Remove the w3m rendering within unpack
The mimetype guessing used in File::Unpack is just super unreliable, this
way we make perfectly readable sources unreadble to save some patterns.

If it turns out that we have tons of html files that all render the GPL
with different css styles interfering, we need to find a different solution.

brute force w3m is none - posibly trying Mojo::Dom and extracting the texts