Obsolete reviews (and their checkouts) more aggressively
So far we kept for every package at least one checksum, but we don't
need that. In contrast we deleted also packages that were in products,
which are reimported the night after.

So simplify the algorithm and just delete all package imports that are
in no product and were (re-)imported a week ago. The 7 days are hard coded
for now - possibly we should make this configurable, but for today it was more
important to save the disk.