Add snippets in an extra table
While indexing, put all snippets with keywords in an extra table.
Retention policy for it to be defined - but we need a place to store
data for the model to learn from and for now we can put the TODO list
there as well. Depends on how large it becomes