Calculate cosinus similarity of patterns
The Levenshtein distance used so far is too expensive too calculate
and most importantely it's impossible to store any intermediate steps.

So switch to cosinus similarity between term frequency–inverse document
frequencies. This way we can store the frequencies of precollected words
in a JSON blob and only calculate the IDF of the document at hand to
then calculate the vector's similarity.

It's still slower than I had expected (it's just a 30K sums of rather
small float vectors, but possibly this is not perl's strength). But <1s
is better than half a minute.