A reimport task got suddenly quite useful over the weekend
Unfortunately it requires more work - as often the infos in bot_sources
are outdated. So for a useful reimport we need to dig for the verifymd5
in all known products. for the weekend incident I did just that as I only
needed to restore those actually in products.

My aim was actually reimporting past reviews, but that is much harder